+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 - Setting `NOT NULL` on an existing column is safe in MySQL 5.6+ and MariaDB 10.0+ with strict mode enabled
 - Added `target_sql_mode` option
 - Added warning for unsupported database versions
-- Changing between `text` and `citext` is safe in Postgres
-- Changing a `varchar` column to a `citext` column is safe in Postgres
-- Changing a `citext` column to a `varchar` column with no length limit is safe in Postgres
+- Changing between `text` and `citext` is safe in Postgres when not indexed
+- Changing a `varchar` column to a `citext` column is safe in Postgres when not indexed
+- Changing a `citext` column to a `varchar` column with no length limit is safe in Postgres when not indexed
 - Changing a `cidr` column to an `inet` column is safe in Postgres
 - Increasing the precision of an `interval` or `time` column is safe in Postgres
 - Decreasing the precision of a `timestamp` or `timestamptz` column is not safe in Postgres